+++ backend/tsconfig.json
@@ -17,7 +17,7 @@
     "skipLibCheck": true,
     "strict": true,
     "target": "esnext",
-    "outDir": "../main"
+    "outDir": "../main/backend"
   },
   "exclude": ["node_modules"],
   "include": ["**/*.ts", "**/*.tsx"]
